+++ config/complete.js
@@ -6,6 +6,8 @@ export const elements = Object.freeze([...HTMLElements, ...SVGElements, ...MathM
 
 export const attributes = Object.freeze(Array.from(new Set([...HTMLAttributes, ...SVGAttributes, ...MathMLAttributes])));
 
-export const comments = false;
+export const comments = true;
 
-export const sanitizer = Object.freeze({ elements, attributes, comments });
+export const dataAttributes = true;
+
+export const sanitizer = Object.freeze({ elements, attributes, comments, dataAttributes });