+++ sanitizer.js
@@ -2,3 +2,4 @@ export { htmlConfig, svgConfig, mathMLConfig, baseConfig, completeConfig } from 
 export { normalizeConfig } from '@aegisjsproject/sanitizer/config-utils.js';
 export { setHTML, parseHTML, sanitize } from '@aegisjsproject/sanitizer/sanitize.js';
 export { HTML as HTMLNS, SVG as SVGNS, MATHML as MATHMLNS } from '@aegisjsproject/sanitizer/namespaces.js';
+export { createSanitizerPolicy } from '@aegisjsproject/sanitizer/trust.js';